Replace old benchmarks with criterion benchmarks
 * cli/Cargo.toml: remove separate crit-benchmark benchmark binary
 * cli/benches/crit-benchmark.rs -> cli/benches/benchmark.rs
 * script/benchmark: pass `--verbose` option to criterion benchmarks for additional statistical data output
 * script/benchmark.cmd: likewise
