+++ cli/benches/benchmark.rs
@@ -1,8 +1,12 @@
+#[macro_use]
+extern crate criterion;
+
+use criterion::{Criterion, Benchmark, Throughput};
 use lazy_static::lazy_static;
 use std::collections::BTreeMap;
 use std::path::{Path, PathBuf};
-use std::time::Instant;
-use std::{env, fs, usize};
+use std::time::Duration;
+use std::{env, fs};
 use tree_sitter::{Language, Parser};
 use tree_sitter_cli::error::Error;
 use tree_sitter_cli::loader::Loader;
@@ -56,126 +60,6 @@ lazy_static! {
     };
 }
 
-fn main() {
-    let mut parser = Parser::new();
-    let max_path_length = EXAMPLE_PATHS_BY_LANGUAGE_DIR
-        .iter()
-        .flat_map(|(_, paths)| paths.iter())
-        .map(|p| p.file_name().unwrap().to_str().unwrap().chars().count())
-        .max()
-        .unwrap();
-
-    let mut all_normal_speeds = Vec::new();
-    let mut all_error_speeds = Vec::new();
-
-    for (language_path, example_paths) in EXAMPLE_PATHS_BY_LANGUAGE_DIR.iter() {
-        let language_name = language_path.file_name().unwrap().to_str().unwrap();
-
-        if let Some(filter) = LANGUAGE_FILTER.as_ref() {
-            if language_name != filter.as_str() {
-                continue;
-            }
-        }
-
-        eprintln!("Benchmarking with {} repetitions", *REPETITION_COUNT);
-
-        eprintln!("\nLanguage: {}", language_name);
-        parser.set_language(get_language(language_path)).unwrap();
-
-        eprintln!("  Normal examples:");
-        let mut normal_speeds = Vec::new();
-        for example_path in example_paths {
-            if let Some(filter) = EXAMPLE_FILTER.as_ref() {
-                if !example_path.to_str().unwrap().contains(filter.as_str()) {
-                    continue;
-                }
-            }
-
-            normal_speeds.push(parse(&mut parser, example_path, max_path_length));
-        }
-
-        eprintln!("  Error examples (mismatched languages):");
-        let mut error_speeds = Vec::new();
-        for (other_language_path, example_paths) in EXAMPLE_PATHS_BY_LANGUAGE_DIR.iter() {
-            if other_language_path != language_path {
-                for example_path in example_paths {
-                    if let Some(filter) = EXAMPLE_FILTER.as_ref() {
-                        if !example_path.to_str().unwrap().contains(filter.as_str()) {
-                            continue;
-                        }
-                    }
-
-                    error_speeds.push(parse(&mut parser, example_path, max_path_length));
-                }
-            }
-        }
-
-        if let Some((average_normal, worst_normal)) = aggregate(&normal_speeds) {
-            eprintln!("  Average Speed (normal): {} bytes/ms", average_normal);
-            eprintln!("  Worst Speed (normal):   {} bytes/ms", worst_normal);
-        }
-
-        if let Some((average_error, worst_error)) = aggregate(&error_speeds) {
-            eprintln!("  Average Speed (errors): {} bytes/ms", average_error);
-            eprintln!("  Worst Speed (errors):   {} bytes/ms", worst_error);
-        }
-
-        all_normal_speeds.extend(normal_speeds);
-        all_error_speeds.extend(error_speeds);
-    }
-
-    eprintln!("\nOverall");
-    if let Some((average_normal, worst_normal)) = aggregate(&all_normal_speeds) {
-        eprintln!("  Average Speed (normal): {} bytes/ms", average_normal);
-        eprintln!("  Worst Speed (normal):   {} bytes/ms", worst_normal);
-    }
-
-    if let Some((average_error, worst_error)) = aggregate(&all_error_speeds) {
-        eprintln!("  Average Speed (errors): {} bytes/ms", average_error);
-        eprintln!("  Worst Speed (errors):   {} bytes/ms", worst_error);
-    }
-    eprintln!("");
-}
-
-fn aggregate(speeds: &Vec<(usize)>) -> Option<(usize, usize)> {
-    if speeds.is_empty() {
-        return None;
-    }
-    let mut total = 0;
-    let mut max = usize::MAX;
-    for speed in speeds.iter().cloned() {
-        total += speed;
-        if speed < max {
-            max = speed;
-        }
-    }
-    Some((total / speeds.len(), max))
-}
-
-fn parse(parser: &mut Parser, example_path: &Path, max_path_length: usize) -> usize {
-    eprint!(
-        "    {:width$}\t",
-        example_path.file_name().unwrap().to_str().unwrap(),
-        width = max_path_length
-    );
-
-    let source_code = fs::read(example_path)
-        .map_err(Error::wrap(|| format!("Failed to read {:?}", example_path)))
-        .unwrap();
-    let time = Instant::now();
-    for _ in 0..*REPETITION_COUNT {
-        parser
-            .parse(&source_code, None)
-            .expect("Incompatible language version");
-    }
-    let duration = time.elapsed() / (*REPETITION_COUNT as u32);
-    let duration_ms =
-        duration.as_secs() as f64 * 1000.0 + duration.subsec_nanos() as f64 / 1000000.0;
-    let speed = (source_code.len() as f64 / duration_ms) as usize;
-    eprintln!("time {} ms\tspeed {} bytes/ms", duration_ms as usize, speed);
-    speed
-}
-
 fn get_language(path: &Path) -> Language {
     let src_dir = GRAMMARS_DIR.join(path).join("src");
     TEST_LOADER
@@ -185,3 +69,104 @@ fn get_language(path: &Path) -> Language {
         }))
         .unwrap()
 }
+
+fn get_normal_inputs(example_paths: &Vec<PathBuf>) -> Vec<(&str, Vec<u8>)> {
+    let mut normal_inputs = Vec::new();
+    for example_path in example_paths {
+        let example_path_str = example_path.to_str().unwrap();
+        let example_file_str = example_path.file_name().unwrap().to_str().unwrap();
+        if let Some(filter) = EXAMPLE_FILTER.as_ref() {
+            if !example_path_str.contains(filter.as_str()) {
+                continue;
+            }
+        }
+        let source_code = fs::read(example_path)
+            .map_err(Error::wrap(|| format!("Failed to read {:?}", example_path)))
+            .unwrap();
+        normal_inputs.push((example_file_str, source_code));
+    }
+    normal_inputs
+}
+
+fn get_error_inputs(language_path: &Path) -> Vec<(&str, Vec<u8>)> {
+    let mut error_inputs = Vec::new();
+    for (other_language_path, example_paths) in EXAMPLE_PATHS_BY_LANGUAGE_DIR.iter() {
+        if other_language_path != language_path {
+            for example_path in example_paths {
+                let example_path_str = example_path.to_str().unwrap();
+                let example_file_str = example_path.file_name().unwrap().to_str().unwrap();
+                if let Some(filter) = EXAMPLE_FILTER.as_ref() {
+                    if !example_path_str.contains(filter.as_str()) {
+                        continue;
+                    }
+                }
+                let source_code = fs::read(example_path)
+                    .map_err(Error::wrap(|| format!("Failed to read {:?}", example_path)))
+                    .unwrap();
+                error_inputs.push((example_file_str, source_code));
+            }
+        }
+    }
+    error_inputs
+}
+
+fn parse(parser: &mut Parser, source_code: &Vec<u8>) {
+    let _tree = parser
+        .parse(&source_code, None)
+        .expect("Incompatible language version");
+}
+
+fn benchmark(c: &mut Criterion) {
+    for (language_path, example_paths) in EXAMPLE_PATHS_BY_LANGUAGE_DIR.iter() {
+        let language_name = language_path.file_name().unwrap().to_str().unwrap();
+
+        if let Some(filter) = LANGUAGE_FILTER.as_ref() {
+            if language_name != filter.as_str() {
+                continue;
+            }
+        }
+
+        let normal_inputs = get_normal_inputs(example_paths);
+        for normal_input in normal_inputs {
+            let mut parser = Parser::new();
+            parser.set_language(get_language(language_path)).unwrap();
+            let file_name = normal_input.0;
+            let source_code = normal_input.1;
+            let source_code_size = source_code.len();
+            c.bench(&format!("language: {}", language_name),
+                    Benchmark::new(format!("parse normal: {}", file_name),
+                                   move |b| b.iter(|| parse(&mut parser, &source_code)),
+                    ).throughput(Throughput::Bytes(source_code_size as u32))
+            );
+        }
+
+        let error_inputs = get_error_inputs(language_path);
+        for error_input in error_inputs {
+            let mut parser = Parser::new();
+            parser.set_language(get_language(language_path)).unwrap();
+            let file_name = error_input.0;
+            let source_code = error_input.1;
+            let source_code_size = source_code.len();
+            c.bench(&format!("language: {}", language_name),
+                    Benchmark::new(format!("parse error: {}", file_name),
+                                   move |b| b.iter(|| parse(&mut parser, &source_code)),
+                    ).throughput(Throughput::Bytes(source_code_size as u32))
+            );
+        }
+    }
+}
+
+fn benchmark_config() -> Criterion {
+    Criterion::default()
+        .warm_up_time(Duration::from_millis(500))
+        .measurement_time(Duration::from_millis(1000))
+        .sample_size(8)
+        .nresamples(1000)
+}
+
+criterion_group!{
+    name = benches;
+    config = benchmark_config();
+    targets = benchmark
+}
+criterion_main!(benches);
